+++ src/main/java/com/troblecodings/signals/blocks/Signal.java
@@ -71,8 +71,8 @@ public class Signal extends BasicBlock {
 
     public static final Map<String, Signal> SIGNALS = new HashMap<>();
     public static final List<Signal> SIGNAL_IDS = new ArrayList<>();
-    public static final PropertyEnum<SignalAngel> ANGEL = PropertyEnum.create("angel",
-            SignalAngel.class);
+    public static final PropertyEnum<SignalAngel> ANGEL =
+            PropertyEnum.create("angel", SignalAngel.class);
     public static final SEProperty CUSTOMNAME = new SEProperty("customname", JsonEnum.BOOLEAN,
             "false", ChangeableStage.AUTOMATICSTAGE, t -> true, 0);
     public static final TileEntitySupplierWrapper SUPPLIER = SignalTileEntity::new;
@@ -122,7 +122,7 @@ public class Signal extends BasicBlock {
     @Override
     public AxisAlignedBB getCollisionBoundingBox(final IBlockState blockState,
             final IBlockAccess worldIn, final BlockPos pos) {
-        return getBoundingBox(blockState, worldIn, pos);
+        return FULL_BLOCK_AABB.expand(20, 10, 20).expand(-20, -10, -20);
     }
 
     public static ItemStack pickBlock(final EntityPlayer player, final Item item) {
@@ -148,8 +148,8 @@ public class Signal extends BasicBlock {
     public IBlockState getStateForPlacement(final World world, final BlockPos pos,
             final EnumFacing facing, final float hitX, final float hitY, final float hitZ,
             final int meta, final EntityLivingBase placer, final EnumHand hand) {
-        final int index = 15
-                - (MathHelper.floor(placer.getRotationYawHead() * 16.0F / 360.0F - 0.5D) & 15);
+        final int index =
+                15 - (MathHelper.floor(placer.getRotationYawHead() * 16.0F / 360.0F - 0.5D) & 15);
         return getDefaultState().withProperty(ANGEL, SignalAngel.values()[index]);
     }